# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: KJIHC Club Management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /ensign:
    get:
      operationId: listEnsignEntries
      tags: [ensign]
      summary: List Ensign Ewart entries
      security: [{ ClerkAuth: [] }]
      parameters:
        - name: ageGroup
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/EnsignEntry"
    post:
      operationId: createEnsignEntry
      tags: [ensign]
      summary: Submit a new Ensign Ewart entry (public)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EnsignInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EnsignEntry"
  /ensign/{id}/payment:
    patch:
      operationId: updateEnsignPayment
      tags: [ensign]
      summary: Update payment status
      security: [{ ClerkAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentStatusUpdate"
      responses:
        "200":
          description: Updated

  # ─── PUBLIC JOIN ──────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    ClerkAuth:
      type: http
      scheme: bearer
  schemas:
...
    EnsignEntry:
      type: object
      required: [id, ageGroup, clubName, teamName, paymentStatus, createdAt]
      properties:
        id:
          type: number
        ageGroup:
          type: string
        clubName:
          type: string
        teamName:
          type: string
        shirtColourHome:
          type: string
        shirtColourAway:
          type: string
        numCoachesOfficials:
          type: string
        seniorContactName:
          type: string
        seniorContactPhoneEmail:
          type: string
        bookingContactName:
          type: string
        bookingContactPhone:
          type: string
        bookingContactEmail:
          type: string
        managerName:
          type: string
        managerPhone:
          type: string
        managerEmail:
          type: string
        remarks:
          type: ["string", "null"]
        paymentStatus:
          type: number
        signature:
          type: string
        printName:
          type: string
        positionInClub:
          type: string
        dateSigned:
          type: string
        createdAt:
          type: string
    EnsignInput:
      type: object
      required: [ageGroup, clubName, teamName, shirtColourHome, shirtColourAway, numCoachesOfficials, seniorContactName, seniorContactPhoneEmail, bookingContactName, bookingContactPhone, bookingContactEmail, managerName, managerPhone, managerEmail, signature, printName, positionInClub, dateSigned]
      properties:
        ageGroup:
          type: string
        clubName:
          type: string
        teamName:
          type: string
        shirtColourHome:
          type: string
        shirtColourAway:
          type: string
        numCoachesOfficials:
          type: string
        seniorContactName:
          type: string
        seniorContactPhoneEmail:
          type: string
        bookingContactName:
          type: string
        bookingContactPhone:
          type: string
        bookingContactEmail:
          type: string
        managerName:
          type: string
        managerPhone:
          type: string
        managerEmail:
          type: string
        remarks:
          type: string
        signature:
          type: string
        printName:
          type: string
        positionInClub:
          type: string
        dateSigned:
          type: string
    PaymentStatusUpdate:
      type: object
      required: [paymentStatus]
      properties:
        paymentStatus:
          type: number

    # ─── JOIN ────────────────────────────────────────────────────────────
...